+++ src/aqua_diagnostics/core/output_saver.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 import os
 import xarray as xr
-from aqua.logger import log_configure
+from aqua.logger import log_configure, log_history
 from aqua.util import create_folder, add_pdf_metadata, add_png_metadata, update_metadata
 
 
@@ -48,8 +48,7 @@ class OutputSaver:
                       catalog_ref: str = None, model_ref: str = None, exp_ref: str = None,
                       extra_keys: dict = None) -> str:
         """
-        Generate a filename based on provided parameters and additional user-defined keywords,
-        including precise time intervals.
+        Generate a filename based on provided parameters and additional user-defined keywords
 
         Args:
             diagnostic_product (str, optional): Product of the diagnostic analysis.
@@ -121,7 +120,7 @@ class OutputSaver:
         self.logger.debug(f"Generated filename: {filename}")
         return filename
 
-    def save_netcdf(self, dataset: xr.Dataset, diagnostic_product: str, rebuild: bool = True, extra_keys: dict = None):
+    def save_netcdf(self, dataset: xr.Dataset, diagnostic_product: str, rebuild: bool = True, extra_keys: dict = None, metadata: dict = None):
         """
         Save an xarray Dataset as a NetCDF file with a generated filename.
 
@@ -130,6 +129,7 @@ class OutputSaver:
             diagnostic_product (str): Product of the diagnostic analysis.
             rebuild (bool, optional): Whether to rebuild the output file if it already exists. Defaults to True.
             extra_keys (dict, optional): Dictionary of additional keys to include in the filename.
+            metadata (dict, optional): Additional metadata to include in the NetCDF file.
         """
         filename = self.generate_name(diagnostic_product=diagnostic_product, extra_keys=extra_keys) + '.nc'
 
@@ -140,6 +140,17 @@ class OutputSaver:
         if not rebuild and os.path.exists(filepath):
             self.logger.info(f"File already exists and rebuild=False, skipping: {filepath}")
             return filepath
+
+        metadata = self.create_metadata(diagnostic_product=diagnostic_product, extra_keys=extra_keys, metadata=metadata)
+       
+        # If metadata contains a history attribute, log the history
+        if 'history' in metadata:
+            log_history(data=dataset, msg=metadata['history'])
+            # Remove the history attribute from the metadata dictionary
+            metadata.pop('history')
+
+        dataset.attrs.update(metadata)
+        
         dataset.to_netcdf(filepath)
 
         self.logger.info(f"Saved NetCDF: {filepath}")